Separate out files to folders, add css files to all components that did not have them, correct file paths to match new file structure
