+++ src/WeatherCard/WeatherCard.js
@@ -1,6 +1,5 @@
 import React from 'react';
-// import $ from 'jquery';
-// import './App.css';
+import './WeatherCard.css';
 
 const WeatherCard = ({ weather }) => {
   let iconUrl = `http://openweathermap.org/img/w/${weather.icon}.png`;
